requirements.txt
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.10.1
Django==2.2.10
django-redis==4.9.0
djangorestframework==3.8.2
drf-yasg==1.7.4
eth-abi==2.1.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
future==0.18.2
geographiclib==1.50
geopy==1.21.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
hexbytes==0.2.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
ipfshttpclient==0.4.12
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lru-dict==1.1.6
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.7.19
openapi-codec==1.3.2
parsimonious==0.8.1
protobuf==3.11.3
pycryptodome==3.9.6
pymongo==3.10.1
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
redis==2.10.6
requests==2.22.0
rlp==1.2.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
toolz==0.10.0
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
varint==1.0.2
web3==5.5.0
websockets==8.1
whitenoise==4.1.2
zipp==2.2.0
